refactor(icons): moving icons to `elements`
