+++ src/components/dropdown/dropdown.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { storiesOf } from '@storybook/react'
 import Dropdown from './dropdown'
-import * as Icon from '../icons'
+import * as Icon from 'elements/icons'
 
 const listItems = [
   {
@@ -27,13 +27,19 @@ const listItems = [
     method: () => {}
   }
 ]
-storiesOf('Components | Dropdown', module).add('active', () => {
-  return (
-    <Dropdown tabId="1" active={true} setStatus={() => {}} list={listItems} />
-  )
-})
-.add('inactive', () => {
-  return (
-    <Dropdown tabId="1" active={false} setStatus={() => {}} list={listItems} />
-  )
-})
+storiesOf('Components | Dropdown', module)
+  .add('active', () => {
+    return (
+      <Dropdown tabId="1" active={true} setStatus={() => {}} list={listItems} />
+    )
+  })
+  .add('inactive', () => {
+    return (
+      <Dropdown
+        tabId="1"
+        active={false}
+        setStatus={() => {}}
+        list={listItems}
+      />
+    )
+  })
