+++ src/components/elements/icons/icons.androidlogo.js
