+++ src/components/elements/icons/icons.arrowright.js
