+++ src/components/elements/icons/icons.close.js
