+++ src/components/elements/icons/icons.cog.js
