+++ src/components/elements/icons/icons.pocketlogotype.js
