+++ src/components/elements/icons/icons.remove.js
