+++ src/components/elements/icons/icons.search.js
