+++ src/components/elements/icons/icons.settings.js
