+++ src/components/elements/icons/icons.shuffle.js
