+++ src/components/elements/icons/icons.sort.js
