+++ src/containers/save/toolbar/toolbar.main.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import Dropdown from 'components/dropdown/dropdown'
 import Tagging from './tagging/tagging'
 import ToolbarError from './toolbar.error'
-import * as Icon from 'components/icons'
+import * as Icon from 'elements/icons'
 import { localize } from 'common/_locales/locales'
 import styled from '@emotion/styled'
 import { COLORS, UTILIZATION } from 'elements/colors/colors'
